cleanups: reorg the output logic under a common module
This is a cleanup effort in prep for removing ns-options.  This
centralizes all of the behavior under a common module in a single
file.  This also improves readability as you get to see all of
the output logic side by side.  It also shows how the outputs
build on the base and all rely on ns-options (for now).

In a future effort, I'll remove ns-options in prep for making this
gem work in modern ruby versions.
